Dev (#5)
* refactor(slave): InterfaceNameIndex pair, uint8_t device_index, readState/writeCommand in base

- Add InterfaceNameIndex = pair<string, size_t> for state/command interfaces
- setInterfaces(device_index, state_interfaces, command_interfaces) with vectors of pair
- readState(state_buffer, set_state) and writeCommand(get_command, cmd_vals) in ModbusSlaveInterface
- device_index type: size_t -> uint8_t in interface and GenericModbusSlave
- Implement setInterfaces/readState/writeCommand in modbus_slave_interface.cpp

Made-with: Cursor

* feat(master): poll thread, MasterParams, state RealtimeBuffer, readStateSnapshotForRT

- Add MasterParams and initFromParams() for connection/poll params parsing
- Poll loop runs inside master; state written to internal RealtimeBuffer
- readStateSnapshotForRT() for RT context to read latest state (no on_state_read callback)
- Add realtime_tools dependency

Made-with: Cursor

* feat: package with examples created files moved from hw if pack

* feat: functional moved to modbus master

* refactoring

* refactoring

* chore: translated

* format

---------

+++ modbus_slave_plugins/include/modbus_slave_plugins/generic_modbus_slave.hpp
@@ -3,19 +3,21 @@
 
 /**
  * @file generic_modbus_slave.hpp
- * @brief Generic Modbus slave plugin: parses device config from YAML (by analogy with GenericEcSlave).
- *        Use in URDF with plugin="modbus_slave_plugins/GenericModbusSlave" and param slave_config or device_config.
+ * @brief Generic Modbus slave plugin: parses device config from YAML (by analogy with
+ * GenericEcSlave). Use in URDF with plugin="modbus_slave_plugins/GenericModbusSlave" and param
+ * slave_config or device_config.
  */
 
 #ifndef MODBUS_SLAVE_PLUGINS__GENERIC_MODBUS_SLAVE_HPP_
 #define MODBUS_SLAVE_PLUGINS__GENERIC_MODBUS_SLAVE_HPP_
 
+#include <cstdint>
 #include <string>
 #include <unordered_map>
 #include <vector>
 
-#include "modbus_slave_plugins/modbus_device_config.hpp"
-#include "modbus_slave_plugins/modbus_slave_interface.hpp"
+#include "modbus_slave_interface/modbus_device_config.hpp"
+#include "modbus_slave_interface/modbus_slave_interface.hpp"
 #include "yaml-cpp/yaml.h"
 
 namespace modbus_slave_plugins {
@@ -23,7 +25,8 @@ namespace modbus_slave_plugins {
 /**
  * Generic Modbus slave plugin.
  * All config parsing is done inside the plugin (setup_from_config_file, setup_from_config),
- * like GenericEcSlave in ethercat_driver_ros2. Hardware interface only loads the plugin and calls setupSlave().
+ * like GenericEcSlave in ethercat_driver_ros2. Hardware interface only loads the plugin and calls
+ * setupSlave().
  */
 class GenericModbusSlave : public modbus_hw_interface::ModbusSlaveInterface {
  public:
@@ -42,14 +45,11 @@ class GenericModbusSlave : public modbus_hw_interface::ModbusSlaveInterface {
   bool setup_from_config_file(const std::string& config_file,
                               modbus_hw_interface::ModbusDeviceConfig& out);
 
-  void updateState(size_t device_index,
-                   const std::vector<std::string>& state_names,
-                   const double* state_values,
-                   size_t count,
+  void updateState(uint8_t device_index, const std::vector<std::string>& state_names,
+                   const double* state_values, size_t count,
                    modbus_hw_interface::SetStateCallback set_state) override;
 
-  void getCommand(size_t device_index,
-                  const std::vector<std::string>& command_names,
+  void getCommand(uint8_t device_index, const std::vector<std::string>& command_names,
                   modbus_hw_interface::GetCommandCallback get_command,
                   std::vector<double>& command_values_out) override;
 };
