+++ modbus_slave_plugins/src/generic_modbus_slave.cpp
@@ -3,7 +3,8 @@
 
 /**
  * @file generic_modbus_slave.cpp
- * @brief Generic Modbus slave plugin: config parsing inside plugin (by analogy with GenericEcSlave).
+ * @brief Generic Modbus slave plugin: config parsing inside plugin (by analogy with
+ * GenericEcSlave).
  */
 
 #include "modbus_slave_plugins/generic_modbus_slave.hpp"
@@ -17,18 +18,18 @@
 
 namespace modbus_slave_plugins {
 
-bool GenericModbusSlave::setupSlave(
-    const std::string& name,
-    const std::unordered_map<std::string, std::string>& parameters,
-    modbus_hw_interface::ModbusDeviceConfig& out) {
+bool GenericModbusSlave::setupSlave(const std::string& name,
+                                    const std::unordered_map<std::string, std::string>& parameters,
+                                    modbus_hw_interface::ModbusDeviceConfig& out) {
   auto it_cfg = parameters.find("device_config");
   if (it_cfg == parameters.end()) {
     it_cfg = parameters.find("slave_config");
   }
   if (it_cfg == parameters.end() || it_cfg->second.empty()) {
-    RCLCPP_ERROR(rclcpp::get_logger("GenericModbusSlave"),
-                 "Modbus slave '%s': missing param 'device_config' or 'slave_config' (path to YAML)",
-                 name.c_str());
+    RCLCPP_ERROR(
+        rclcpp::get_logger("GenericModbusSlave"),
+        "Modbus slave '%s': missing param 'device_config' or 'slave_config' (path to YAML)",
+        name.c_str());
     return false;
   }
 
@@ -45,8 +46,8 @@ bool GenericModbusSlave::setupSlave(
 
   if (!setup_from_config_file(it_cfg->second, out)) {
     RCLCPP_ERROR(rclcpp::get_logger("GenericModbusSlave"),
-                 "Failed to load device config from '%s' for slave '%s'",
-                 it_cfg->second.c_str(), name.c_str());
+                 "Failed to load device config from '%s' for slave '%s'", it_cfg->second.c_str(),
+                 name.c_str());
     return false;
   }
   out.name = name;  // ensure name from URDF overrides any in YAML
@@ -60,8 +61,7 @@ bool GenericModbusSlave::setup_from_config(const YAML::Node& slave_config,
                  "GenericModbusSlave: failed to load slave configuration: empty configuration");
     return false;
   }
-  modbus_hw_interface::ModbusDeviceConfigLoader loader(
-      rclcpp::get_logger("GenericModbusSlave"));
+  modbus_hw_interface::ModbusDeviceConfigLoader loader(rclcpp::get_logger("GenericModbusSlave"));
   return loader.loadFromNode("slave_config", slave_config, out);
 }
 
@@ -82,20 +82,19 @@ bool GenericModbusSlave::setup_from_config_file(const std::string& config_file,
   }
 }
 
-void GenericModbusSlave::updateState(size_t /*device_index*/,
-                                    const std::vector<std::string>& state_names,
-                                    const double* state_values,
-                                    size_t count,
-                                    modbus_hw_interface::SetStateCallback set_state) {
+void GenericModbusSlave::updateState(uint8_t /*device_index*/,
+                                     const std::vector<std::string>& state_names,
+                                     const double* state_values, size_t count,
+                                     modbus_hw_interface::SetStateCallback set_state) {
   for (size_t i = 0; i < count && i < state_names.size(); ++i) {
     set_state(state_names[i], state_values[i]);
   }
 }
 
-void GenericModbusSlave::getCommand(size_t /*device_index*/,
-                                   const std::vector<std::string>& command_names,
-                                   modbus_hw_interface::GetCommandCallback get_command,
-                                   std::vector<double>& command_values_out) {
+void GenericModbusSlave::getCommand(uint8_t /*device_index*/,
+                                    const std::vector<std::string>& command_names,
+                                    modbus_hw_interface::GetCommandCallback get_command,
+                                    std::vector<double>& command_values_out) {
   command_values_out.resize(command_names.size());
   for (size_t i = 0; i < command_names.size(); ++i) {
     command_values_out[i] = get_command(command_names[i]);
